+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 * Add `StreamExtensions.firstOrNull`.
 
+* Add a `CancelableOperation.fromSubscription()` static factory.
+
+* Add a `CancelableOperation.race()` static method.
+
 ## 2.8.2
 
 * Deprecate `EventSinkBase`, `StreamSinkBase`, `IOSinkBase`.